moved photogrid and added simple test for app.
