+++ src/components/PhotoGrid.tsx
